requirements.txt
azure-functions
azure-functions-durable
azure-storage-blob
google-generativeai>=0.3.0
anthropic>=0.18.0
requests
regex
tenacity
python-docx>=0.8.11
pdfplumber>=0.9.0